add usesInvoice condition to hide link, convert to ts and add test
