--- src/pages/MembersPage/MembersActivation/Activation/ChangePlanLink/index.js
@@ -1,1 +1,0 @@
-export { default } from './ChangePlanLink'
